most configuration moved to 'after-init-hook
When packages are not installed on a host, I usually get in a
back-and-forth with using `emacs --debug-init` to prevent emacs from
running customization code, but then we do not have our
package-selected-packages initialized, so cannot refresh package
contents and install missing packages.

The point of this change is to move all customization that could fail
due to missing packages into a callback that only gets invoked after
packages have been loaded.

More work remains to be done. I would like to enumerate over the
package-selected-packages, install each one if not already installed,
and then continue with initialization.